+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
-# Changelog for OHLC v1.1.6
+# Changelog for OHLC v1.1.7
 
-Typespec fix.
+Fixed the weekly candles stime and etime.
+Testing refactoring.